% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_arm_cpp}
\alias{generate_arm_cpp}
\title{Generate patient-level data for one treatment group (internal)}
\usage{
generate_arm_cpp(
  nsim,
  n,
  lam_p,
  theta,
  c_resp,
  z_tau,
  tau,
  os_model,
  pi_death,
  gam0,
  gam1,
  lam_o,
  c_dec,
  grid_g,
  grid_h,
  ttr_mode,
  ttr_rate,
  a_time,
  a_cum,
  d_hazard
)
}
\arguments{
\item{nsim}{Number of simulated trials.}

\item{n}{Number of patients per trial in this group.}

\item{lam_p, theta, c_resp, z_tau, tau}{Parameters of PFS and response.}

\item{os_model}{0 for "idm", 1 for "expexp".}

\item{pi_death, gam0, gam1}{Parameters of the illness-death model.}

\item{lam_o, c_dec, grid_g, grid_h}{Parameters of the exponential-exponential
model and its tabulated cumulative post-progression hazard.}

\item{ttr_mode}{1 to generate the time to response.}

\item{ttr_rate}{Rate of the untruncated time to response.}

\item{a_time, a_cum}{Accrual breakpoints and cumulative probabilities.}

\item{d_hazard}{Dropout hazard (0 for no dropout).}
}
\value{
A list of vectors.
}
\description{
Internal C++ kernel of \code{\link{rOncoEndpoints}}. Random numbers are
drawn from the \code{dqrng} generator in the following order, each as one
block of \code{nsim * n} values: latent PFS score, independent normal for
response, uniform for the type of the PFS event, standard exponential for
post-progression survival, uniform for the time to response, uniform for
accrual, standard exponential for dropout. The draws are then transformed
by \code{coronco::transform_one()} (see \code{transform_core.h}).
}
\keyword{internal}
