% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.OncoArm.R
\name{print.OncoArm}
\alias{print.OncoArm}
\title{Print an OncoArm object}
\usage{
\method{print}{OncoArm}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{OncoArm}.}

\item{digits}{Number of significant digits.}

\item{...}{Not used.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints the calibrated parameters of a treatment group together with the
implied medians of PFS and OS (overall and by response) and the implied
Pearson correlations among PFS, OS and response.
}
\examples{
OncoArm(pfs.median = 6, orr = 0.3, resp.cor = 0.4, death.prop = 0.15,
        os.median = 15)
}
