% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuantileEndpoint.R
\name{QuantileEndpoint}
\alias{QuantileEndpoint}
\title{Quantiles of PFS or OS}
\usage{
QuantileEndpoint(
  arm,
  prob = 0.5,
  endpoint = c("os", "pfs"),
  response = c("all", "responders", "nonresponders")
)
}
\arguments{
\item{arm}{An object of class \code{OncoArm}.}

\item{prob}{Numeric vector of probabilities in (0, 1).}

\item{endpoint}{\code{"os"} or \code{"pfs"}.}

\item{response}{\code{"all"}, \code{"responders"} or \code{"nonresponders"}.}
}
\value{
Numeric vector of times.
}
\description{
Computes the time at which the survival function of PFS or OS, for all
patients or by response, equals \code{1 - prob}. The median is
\code{prob = 0.5}.
}
\examples{
arm <- OncoArm(pfs.median = 6, orr = 0.3, resp.cor = 0.4, death.prop = 0.15,
               os.median = 15)
QuantileEndpoint(arm, endpoint = "os")
QuantileEndpoint(arm, endpoint = "pfs", response = "responders")
}
